+++ src/admin/__tests__/deleteTopicRecords.spec.js
@@ -338,10 +338,10 @@ describe('Admin > deleteTopicRecords', () => {
     }
 
     expect(error).toBeDefined()
-    expect(error.name).toBe('KafkaJSNumberOfRetriesExceeded')
-    expect(error.originalError.name).toBe('KafkaJSDeleteTopicRecordsError')
-    expect(error.originalError.retriable).toBe(false)
-    expect(error.originalError.partitions).toEqual([
+    expect(error.name).toBe('KafkaJSNonRetriableError')
+    expect(error.cause.name).toBe('KafkaJSDeleteTopicRecordsError')
+    expect(error.cause.retriable).toBe(false)
+    expect(error.cause.partitions).toEqual([
       {
         partition: 1,
         offset: '99',
@@ -402,10 +402,10 @@ describe('Admin > deleteTopicRecords', () => {
     }
 
     expect(error).toBeDefined()
-    expect(error.name).toBe('KafkaJSNumberOfRetriesExceeded')
-    expect(error.originalError.name).toBe('KafkaJSDeleteTopicRecordsError')
-    expect(error.originalError.retriable).toBe(false)
-    expect(error.originalError.partitions).toEqual(
+    expect(error.name).toBe('KafkaJSNonRetriableError')
+    expect(error.cause.name).toBe('KafkaJSDeleteTopicRecordsError')
+    expect(error.cause.retriable).toBe(false)
+    expect(error.cause.partitions).toEqual(
       expect.arrayContaining([
         {
           partition: 0,